openscad/cad_project/requirements.txt
cadquery>=2.4
pytest>=7.0
trimesh>=4.0
pillow>=10.0
numpy>=1.24
